requirements.txt
python-telegram-bot==20.6
pylibdmtx==0.1.10
Pillow==10.1.0
